requirements.txt
django>=1.6
django-fsm>=2.0.1
django-fsm-admin>=1.0.0
